ref(dynamic-sampling): Cleanup tasks for dynamic sampling (#50721)
