+++ pyproject.toml
@@ -499,13 +499,6 @@ module = [
     "sentry.discover.endpoints.discover_saved_queries",
     "sentry.discover.endpoints.serializers",
     "sentry.discover.tasks",
-    "sentry.dynamic_sampling.prioritise_projects",
-    "sentry.dynamic_sampling.prioritise_transactions",
-    "sentry.dynamic_sampling.recalibrate_transactions",
-    "sentry.dynamic_sampling.rules.base",
-    "sentry.dynamic_sampling.sliding_window",
-    "sentry.dynamic_sampling.snuba_utils",
-    "sentry.dynamic_sampling.tasks",
     "sentry.event_manager",
     "sentry.eventstore.base",
     "sentry.eventstore.compressor",
@@ -1386,16 +1379,6 @@ module = [
     "tests.sentry.digests.backends.test_redis",
     "tests.sentry.digests.test_notifications",
     "tests.sentry.digests.test_utilities",
-    "tests.sentry.dynamic_sampling.models.test_projects_rebalancing",
-    "tests.sentry.dynamic_sampling.models.test_transactions_rebalancing",
-    "tests.sentry.dynamic_sampling.rules.biases.test_boost_latest_releases_bias",
-    "tests.sentry.dynamic_sampling.test_generate_rules",
-    "tests.sentry.dynamic_sampling.test_logging",
-    "tests.sentry.dynamic_sampling.test_prioritise_projects",
-    "tests.sentry.dynamic_sampling.test_prioritise_transactions",
-    "tests.sentry.dynamic_sampling.test_recalibrate_transactions",
-    "tests.sentry.dynamic_sampling.test_sliding_window",
-    "tests.sentry.dynamic_sampling.test_tasks",
     "tests.sentry.event_manager.interfaces.test_breadcrumbs",
     "tests.sentry.event_manager.interfaces.test_contexts",
     "tests.sentry.event_manager.interfaces.test_csp",
