+++ src/sentry/dynamic_sampling/tasks/common.py
@@ -1,0 +1,187 @@
+import logging
+import math
+import time
+from datetime import datetime, timedelta
+from typing import Generator, List, Optional, Tuple
+
+import sentry_sdk
+from snuba_sdk import (
+    Column,
+    Condition,
+    Direction,
+    Entity,
+    Function,
+    Granularity,
+    Op,
+    OrderBy,
+    Query,
+    Request,
+)
+
+from sentry import quotas
+from sentry.dynamic_sampling.rules.helpers.sliding_window import extrapolate_monthly_volume
+from sentry.dynamic_sampling.tasks.constants import (
+    CHUNK_SIZE,
+    MAX_ORGS_PER_QUERY,
+    MAX_PROJECTS_PER_QUERY,
+    MAX_SECONDS,
+)
+from sentry.dynamic_sampling.tasks.logging import log_extrapolated_monthly_volume
+from sentry.sentry_metrics import indexer
+from sentry.snuba.dataset import Dataset, EntityKey
+from sentry.snuba.metrics.naming_layer.mri import TransactionMRI
+from sentry.snuba.referrer import Referrer
+from sentry.utils.snuba import raw_snql_query
+
+logger = logging.getLogger(__name__)
+
+
+def get_active_orgs_with_projects_counts(
+    max_orgs: int = MAX_ORGS_PER_QUERY, max_projects: int = MAX_PROJECTS_PER_QUERY
+) -> Generator[List[int], None, None]:
+    """
+    Fetch organisations in batches.
+    A batch will return at max max_orgs elements
+    It will accumulate org ids in the list until either it accumulates max_orgs or the
+    number of projects in the already accumulated orgs is more than max_projects or there
+    are no more orgs
+    """
+    start_time = time.time()
+    metric_id = indexer.resolve_shared_org(str(TransactionMRI.COUNT_PER_ROOT_PROJECT.value))
+    offset = 0
+    last_result: List[Tuple[int, int]] = []
+    while (time.time() - start_time) < MAX_SECONDS:
+        query = (
+            Query(
+                match=Entity(EntityKey.GenericOrgMetricsCounters.value),
+                select=[
+                    Function("uniq", [Column("project_id")], "num_projects"),
+                    Column("org_id"),
+                ],
+                groupby=[
+                    Column("org_id"),
+                ],
+                where=[
+                    Condition(Column("timestamp"), Op.GTE, datetime.utcnow() - timedelta(hours=1)),
+                    Condition(Column("timestamp"), Op.LT, datetime.utcnow()),
+                    Condition(Column("metric_id"), Op.EQ, metric_id),
+                ],
+                granularity=Granularity(3600),
+                orderby=[
+                    OrderBy(Column("org_id"), Direction.ASC),
+                ],
+            )
+            .set_limit(CHUNK_SIZE + 1)
+            .set_offset(offset)
+        )
+        request = Request(
+            dataset=Dataset.PerformanceMetrics.value, app_id="dynamic_sampling", query=query
+        )
+        data = raw_snql_query(
+            request,
+            referrer=Referrer.DYNAMIC_SAMPLING_COUNTERS_FETCH_PROJECTS_WITH_COUNT_PER_TRANSACTION.value,  # type:ignore
+        )["data"]
+        count = len(data)
+        more_results = count > CHUNK_SIZE
+        offset += CHUNK_SIZE
+        if more_results:
+            data = data[:-1]
+        for row in data:
+            last_result.append((row["org_id"], row["num_projects"]))
+
+        first_idx = 0
+        count_projects = 0
+        for idx, (org_id, num_projects) in enumerate(last_result):
+            count_projects += num_projects
+            if idx - first_idx >= max_orgs - 1 or count_projects >= max_projects:
+                # we got to the number of elements desired
+                yield [o for o, _ in last_result[first_idx : idx + 1]]
+                first_idx = idx + 1
+                count_projects = 0
+
+        # keep what is left unused from last_result for the next iteration or final result
+        last_result = last_result[first_idx:]
+        if not more_results:
+            break
+    if len(last_result) > 0:
+        yield [org_id for org_id, _ in last_result]
+
+
+def sample_rate_to_float(sample_rate: Optional[str]) -> Optional[float]:
+    """
+    Converts a sample rate to a float or returns None in case the conversion failed.
+    """
+    if sample_rate is None:
+        return None
+
+    try:
+        return float(sample_rate)
+    except (TypeError, ValueError):
+        return None
+
+
+def are_equal_with_epsilon(a: Optional[float], b: Optional[float]) -> bool:
+    """
+    Checks if two floating point numbers are equal within an error boundary.
+    """
+    if a is None and b is None:
+        return True
+
+    if a is None or b is None:
+        return False
+
+    return math.isclose(a, b)
+
+
+def compute_guarded_sliding_window_sample_rate(
+    org_id: int, project_id: Optional[int], total_root_count: int, window_size: int
+) -> Optional[float]:
+    """
+    Computes the actual sliding window sample rate by guarding any exceptions and returning None in case
+    any problem would arise.
+    """
+    try:
+        # We want to compute the sliding window sample rate by considering a window of time.
+        # This piece of code is very delicate, thus we want to guard it properly and capture any errors.
+        return compute_sliding_window_sample_rate(org_id, project_id, total_root_count, window_size)
+    except Exception as e:
+        sentry_sdk.capture_exception(e)
+        return None
+
+
+def compute_sliding_window_sample_rate(
+    org_id: int, project_id: Optional[int], total_root_count: int, window_size: int
+) -> Optional[float]:
+    """
+    Computes the actual sample rate for the sliding window given the total root count and the size of the
+    window that was used for computing the root count.
+
+    The org_id is used only because it is required on the quotas side to determine whether dynamic sampling is
+    enabled in the first place for that project.
+    """
+    extrapolated_volume = extrapolate_monthly_volume(volume=total_root_count, hours=window_size)
+    if extrapolated_volume is None:
+        with sentry_sdk.push_scope() as scope:
+            scope.set_extra("org_id", org_id)
+            scope.set_extra("window_size", window_size)
+            sentry_sdk.capture_message("The volume of the current month can't be extrapolated.")
+
+        return None
+
+    # We want to log the monthly volume for observability purposes.
+    log_extrapolated_monthly_volume(
+        org_id, project_id, total_root_count, extrapolated_volume, window_size
+    )
+
+    sampling_tier = quotas.get_transaction_sampling_tier_for_volume(  # type:ignore
+        org_id, extrapolated_volume
+    )
+    if sampling_tier is None:
+        return None
+
+    # We unpack the tuple containing the sampling tier information in the form (volume, sample_rate). This is done
+    # under the assumption that the sampling_tier tuple contains both non-null values.
+    _, sample_rate = sampling_tier
+
+    # We assume that the sample_rate is a float.
+    return float(sample_rate)
