+++ tests/sentry/dynamic_sampling/tasks/test_boost_low_volume_transactions.py
@@ -1,9 +1,12 @@
-from datetime import datetime, timezone
+from datetime import timedelta
+from typing import Tuple
 
+from django.utils import timezone
 from freezegun import freeze_time
 
-from sentry.dynamic_sampling.prioritise_transactions import (
+from sentry.dynamic_sampling.tasks.boost_low_volume_transactions import (
     ProjectIdentity,
+    ProjectTransactions,
     ProjectTransactionsTotals,
     fetch_project_transaction_totals,
     fetch_transactions_with_total_volumes,
@@ -14,10 +17,13 @@ from sentry.dynamic_sampling.prioritise_transactions import (
     next_totals,
     transactions_zip,
 )
-from sentry.snuba.metrics import TransactionMRI
+from sentry.dynamic_sampling.tasks.recalibrate_orgs import fetch_org_volumes, get_active_orgs
+from sentry.snuba.metrics.naming_layer.mri import TransactionMRI
 from sentry.testutils import BaseMetricsLayerTestCase, SnubaTestCase, TestCase
 
-MOCK_DATETIME = datetime(2023, 8, 7, 0, 0, 0, tzinfo=timezone.utc)
+MOCK_DATETIME = (timezone.now() - timedelta(days=1)).replace(
+    hour=0, minute=0, second=0, microsecond=0
+)
 
 
 @freeze_time(MOCK_DATETIME)
@@ -139,21 +145,26 @@ class PrioritiseProjectsSnubaQueryTest(BaseMetricsLayerTestCase, TestCase, Snuba
 
 
 def test_merge_transactions_full():
-    t1 = {
+    t1: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("ts1", 10), ("tm2", 100)],
         "total_num_transactions": None,
         "total_num_classes": None,
     }
-    t2 = {
+    t2: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("tm2", 100), ("tl3", 1000)],
         "total_num_transactions": None,
         "total_num_classes": None,
     }
-    counts = {"project_id": 1, "org_id": 2, "total_num_transactions": 5555, "total_num_classes": 20}
+    counts: ProjectTransactionsTotals = {
+        "project_id": 1,
+        "org_id": 2,
+        "total_num_transactions": 5555,
+        "total_num_classes": 20,
+    }
     actual = merge_transactions(t1, t2, counts)
 
     expected = {
@@ -168,14 +179,14 @@ def test_merge_transactions_full():
 
 
 def test_merge_transactions_missing_totals():
-    t1 = {
+    t1: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("ts1", 10), ("tm2", 100)],
         "total_num_transactions": None,
         "total_num_classes": None,
     }
-    t2 = {
+    t2: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("tm2", 100), ("tl3", 1000)],
@@ -185,7 +196,7 @@ def test_merge_transactions_missing_totals():
 
     actual = merge_transactions(t1, t2, None)
 
-    expected = {
+    expected: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("ts1", 10), ("tm2", 100), ("tl3", 1000)],
@@ -197,17 +208,22 @@ def test_merge_transactions_missing_totals():
 
 
 def test_merge_transactions_missing_right():
-    t1 = {
+    t1: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("ts1", 10), ("tm2", 100)],
         "total_num_transactions": None,
         "total_num_classes": None,
     }
-    counts = {"project_id": 1, "org_id": 2, "total_num_transactions": 5555, "total_num_classes": 20}
+    counts: ProjectTransactionsTotals = {
+        "project_id": 1,
+        "org_id": 2,
+        "total_num_transactions": 5555,
+        "total_num_classes": 20,
+    }
     actual = merge_transactions(t1, None, counts)
 
-    expected = {
+    expected: ProjectTransactions = {
         "project_id": 1,
         "org_id": 2,
         "transaction_counts": [("ts1", 10), ("tm2", 100)],
@@ -269,11 +285,11 @@ def test_transactions_zip():
 
 
 def test_same_project():
-    p1 = {"project_id": 1, "org_id": 2}
-    p1bis = {"project_id": 1, "org_id": 2}
-    p2 = {"project_id": 1, "org_id": 3}
-    p3 = {"project_id": 2, "org_id": 1}
-    p4 = {"project_id": 3, "org_id": 4}
+    p1: ProjectIdentity = {"project_id": 1, "org_id": 2}
+    p1bis: ProjectIdentity = {"project_id": 1, "org_id": 2}
+    p2: ProjectIdentity = {"project_id": 1, "org_id": 3}
+    p3: ProjectIdentity = {"project_id": 2, "org_id": 1}
+    p4: ProjectIdentity = {"project_id": 3, "org_id": 4}
 
     assert is_same_project(p1, p1bis)
     assert not is_same_project(p1, p2)
@@ -282,11 +298,11 @@ def test_same_project():
 
 
 def test_project_before():
-    p1 = {"project_id": 1, "org_id": 2}
-    p1bis = {"project_id": 1, "org_id": 2}
-    p2 = {"project_id": 1, "org_id": 3}
-    p3 = {"project_id": 2, "org_id": 2}
-    p4 = {"project_id": 2, "org_id": 1}
+    p1: ProjectIdentity = {"project_id": 1, "org_id": 2}
+    p1bis: ProjectIdentity = {"project_id": 1, "org_id": 2}
+    p2: ProjectIdentity = {"project_id": 1, "org_id": 3}
+    p3: ProjectIdentity = {"project_id": 2, "org_id": 2}
+    p4: ProjectIdentity = {"project_id": 2, "org_id": 1}
 
     # same project
     assert not is_project_identity_before(p1, p1bis)
@@ -348,3 +364,84 @@ def test_next_totals():
     assert get_totals(pi(3, 1)) is None
     assert get_totals(pi(3, 2)) is None
     assert get_totals(pi(3, 3)) is None
+
+
+@freeze_time(MOCK_DATETIME)
+class FetchOrgVolumesSnubaQueryTest(BaseMetricsLayerTestCase, TestCase, SnubaTestCase):
+    @property
+    def now(self):
+        return MOCK_DATETIME
+
+    def setUp(self):
+        super().setUp()
+        self.orgs_info = []
+        self.num_orgs = 3
+        self.num_proj = 2
+        self.num_trans = 2
+        # create some orgs, projects and transactions
+        for org_idx in range(self.num_orgs):
+            org = self.create_organization(f"test-org{org_idx}")
+            org_info = {"org_id": org.id, "project_ids": []}
+            self.orgs_info.append(org_info)
+            for proj_idx in range(self.num_proj):
+                p = self.create_project(organization=org)
+                org_info["project_ids"].append(p.id)
+                for name_idx in range(1, self.num_trans + 1):
+                    # make up some unique count
+                    count_transactions_keep = name_idx + proj_idx * 10
+                    count_transactions_drop = count_transactions_keep + org_idx * 100
+
+                    self.store_performance_metric(
+                        name=TransactionMRI.COUNT_PER_ROOT_PROJECT.value,
+                        tags={"transaction": f"t-{name_idx}", "decision": "drop"},
+                        minutes_before_now=2,
+                        value=count_transactions_drop,
+                        project_id=p.id,
+                        org_id=org.id,
+                    )
+                    self.store_performance_metric(
+                        name=TransactionMRI.COUNT_PER_ROOT_PROJECT.value,
+                        tags={"transaction": f"t-{name_idx}", "decision": "keep"},
+                        minutes_before_now=2,
+                        value=count_transactions_keep,
+                        project_id=p.id,
+                        org_id=org.id,
+                    )
+        self.org_ids = [org["org_id"] for org in self.orgs_info]
+
+    def get_counts_for_org(self, org_id: int) -> Tuple[int, int]:
+        for org_idx, cur_org_id in enumerate(self.org_ids):
+            if cur_org_id == org_id:
+                break
+        else:
+            raise ValueError("Invalid organisation passed")
+
+        trans = self.num_trans
+        projs = self.num_proj
+
+        # calculate how many transactions we created in the setup for this org
+        total_keep = int(trans * (trans + 1) / 2) * projs + 10 * (projs - 1) * trans
+        total_drop = total_keep + 100 * org_idx * trans * projs
+
+        return total_drop + total_keep, total_keep
+
+    def test_fetch_org_volumes(self):
+        org_counts = fetch_org_volumes(self.org_ids, timedelta(minutes=5))
+
+        for actual_counts in org_counts:
+            current_org = actual_counts.org_id
+            expected_total, expected_keep = self.get_counts_for_org(current_org)
+
+            assert actual_counts.total == expected_total
+            assert actual_counts.indexed == expected_keep
+
+    def test_get_active_orgs(self):
+        all_orgs = []
+        for orgs in get_active_orgs(2, time_interval=timedelta(minutes=5)):
+            all_orgs += orgs
+
+        # test we return the expected number of unique orgs
+        assert len(set(all_orgs)) == len(self.org_ids)
+
+        for org in self.org_ids:
+            assert org in self.org_ids
