+++ tests/sentry/dynamic_sampling/tasks/test_sliding_window.py
@@ -3,11 +3,13 @@ from datetime import timedelta
 from django.utils import timezone
 from freezegun import freeze_time
 
-from sentry.dynamic_sampling.sliding_window import (
-    fetch_orgs_with_total_root_transactions_count,
+from sentry.dynamic_sampling.tasks.sliding_window import (
     fetch_projects_with_total_root_transactions_count,
 )
-from sentry.snuba.metrics import TransactionMRI
+from sentry.dynamic_sampling.tasks.sliding_window_org import (
+    fetch_orgs_with_total_root_transactions_count,
+)
+from sentry.snuba.metrics.naming_layer.mri import TransactionMRI
 from sentry.testutils import BaseMetricsLayerTestCase, SnubaTestCase, TestCase
 
 MOCK_DATETIME = (timezone.now() - timedelta(days=1)).replace(
